Update ingress to use networking.k8s.io/v1 API
This enables the controller to run on Kubernetes v1.22, which has removed
ingress from extensions/v1beta1.

This updates the client-go library to v0.19.x, which introduced ingress to
networking/v1. Updating to v0.22.x is a more involved change as it requires
updates to a number of other dependencies.